+++ src/commands/run/index.ts
@@ -17,6 +17,10 @@ export const options: CommandOptions = [
 	{
 		flags: '-c, --config <path to file>',
 		description: 'define configurations for a test profile'
+	},
+	{
+		flags: '-r, --recreate',
+		description: 'recreate the project'
 	}
 ];
 
@@ -37,14 +41,18 @@ export const action = async (params: Command) => {
 	await execs.sleep(1000);
 
 	execs.display('=> Cleaning old data.'.blue);
-	execs.execute('docker system prune -a --volumes');
+	if (currentOptions.recreate) {
+		execs.execute('docker system prune -a --volumes');
+	}
 
 	execs.display('\nStarting project.');
 	for (const server of result.servers) {
 		execs.display(`=> Starting '${server.name}'.`.blue);
 
 		const server_path = path.join(result.output.absolute, server.name, server.file);
-		execs.execute(`docker-compose -f ${server_path} down -v`)
+		if (currentOptions.recreate) {
+			execs.execute(`docker-compose -f ${server_path} down -v`);
+		}
 		execs.execute(`docker-compose -f ${server_path} up --detach`);
 	}
